+++ xenon-common/src/main/java/com/vmware/xenon/services/common/RoleService.java
@@ -161,16 +161,13 @@ public class RoleService extends StatefulService {
 
         RoleState currentState = getState(op);
         ServiceDocumentDescription documentDescription = getStateDescription();
+
         if (ServiceDocument.equals(documentDescription, currentState, newState)) {
-            // HTTP-304 spec doesn't define behavior for PUT
-            // In current implementation, setting 304 will skip creating new version document in index.
-            // This is not appropriate behavior from http standpoint, and it may change in future.
-            // (Probably use other means such as pragma to skip creating new version)
-            // The behavior for response operation is also relies on current xenon implementation and may change in future.
-            op.setStatusCode(Operation.STATUS_CODE_NOT_MODIFIED);
-        } else {
-            setState(op, newState);
+            // do not update state
+            op.addPragmaDirective(Operation.PRAGMA_DIRECTIVE_STATE_NOT_MODIFIED);
         }
+
+        setState(op, newState);
         op.complete();
     }