+++ xenon-common/src/test/java/com/vmware/xenon/services/common/TestRoleService.java
@@ -19,7 +19,6 @@ import static org.junit.Assert.assertTrue;
 import java.net.URI;
 import java.util.HashSet;
 import java.util.Set;
-import java.util.UUID;
 import java.util.function.Supplier;
 
 import org.junit.After;
@@ -87,9 +86,24 @@ public class TestRoleService extends BasicReusableHostTestCase {
     @Test
     public void testFactoryIdempotentPost() throws Throwable {
         RoleState state = validRoleState();
-        state.documentSelfLink = UUID.randomUUID().toString();
+        String servicePath = UriUtils.buildUriPath(RoleService.FACTORY_LINK, "my-role");
+        state.documentSelfLink = servicePath;
 
-        RoleState responseState = (RoleState) this.host.verifyPost(RoleState.class,
+        RoleState responseState = this.host.verifyPost(RoleState.class,
+                ServiceUriPaths.CORE_AUTHZ_ROLES,
+                state,
+                Operation.STATUS_CODE_OK);
+
+        assertEquals(state.userGroupLink, responseState.userGroupLink);
+        assertEquals(state.resourceGroupLink, responseState.resourceGroupLink);
+        assertEquals(state.verbs, responseState.verbs);
+        assertEquals(state.priority, responseState.priority);
+        assertEquals(state.policy, responseState.policy);
+        long initialVersion = responseState.documentVersion;
+
+        // second post should be converted to put
+        // Since sending same document, this post/put should not persist(increment) the document
+        responseState = this.host.verifyPost(RoleState.class,
                 ServiceUriPaths.CORE_AUTHZ_ROLES,
                 state,
                 Operation.STATUS_CODE_OK);
@@ -100,14 +114,12 @@ public class TestRoleService extends BasicReusableHostTestCase {
         assertEquals(state.priority, responseState.priority);
         assertEquals(state.policy, responseState.policy);
 
-        this.host.verifyPost(RoleState.class,
-                ServiceUriPaths.CORE_AUTHZ_ROLES,
-                state,
-                Operation.STATUS_CODE_NOT_MODIFIED);
+        RoleState getState = this.sender.sendAndWait(Operation.createGet(this.host, servicePath), RoleState.class);
+        assertEquals("version should not increase", initialVersion, getState.documentVersion);
 
+        // modify document
         state.verbs.add(Action.PATCH);
-
-        responseState = (RoleState) this.host.verifyPost(RoleState.class,
+        responseState = this.host.verifyPost(RoleState.class,
                 ServiceUriPaths.CORE_AUTHZ_ROLES,
                 state,
                 Operation.STATUS_CODE_OK);
@@ -117,6 +129,7 @@ public class TestRoleService extends BasicReusableHostTestCase {
         assertEquals(state.verbs, responseState.verbs);
         assertEquals(state.priority, responseState.priority);
         assertEquals(state.policy, responseState.policy);
+        assertTrue("version should increase", initialVersion < responseState.documentVersion);
     }
 
     void testFactoryPostFailure(Supplier<RoleState> sup) throws Throwable {